+++ src/serialize.rs
@@ -45,7 +45,7 @@ pub fn serialize<T: serde::Serialize + ?Sized, W: CoreWrite, O: Options>(
 pub fn serialize_size<T: serde::Serialize + ?Sized, O: Options>(
     value: &T,
     options: O,
-) -> Result<u64, SerializeError<()>> {
+) -> Result<usize, SerializeError<()>> {
     let mut size_checker = crate::size_checker::SizeChecker { options, total: 0 };
     value.serialize(&mut size_checker)?;
     Ok(size_checker.total)
